+++ scripts/wdio/clean-screenshots.js
@@ -1,26 +1,20 @@
 const fs = require('fs-extra');
 const glob = require('glob');
-const path = require('path');
 const Logger = require('../utils/logger');
 
 // eslint-disable-next-line global-require, import/no-dynamic-require
-const loadWdioConfig = configPath => require(path.resolve(configPath));
-
 const isDirectory = filePath => (fs.existsSync(filePath) && fs.lstatSync(filePath).isDirectory());
 
 const cleanSnapshots = (options) => {
   const {
-    configPath, removeReference,
+    removeReference,
   } = options;
-
-  const wdioConfig = loadWdioConfig(configPath).config;
-  const errorDir = wdioConfig.screenshotPath || '';
 
   const patterns = [
     `${process.cwd()}/**/__snapshots__/latest`,
     `${process.cwd()}/**/__snapshots__/diff`,
     `${process.cwd()}/**/__snapshots__/screen`,
-    errorDir,
+    `${process.cwd()}/errorScreenshots`,
   ];
 
   if (removeReference) {
@@ -41,9 +35,6 @@ const cleanSnapshots = (options) => {
   });
 
   Logger.log('Cleaned screenshot directories\n', { context: '[Terra-Tookit:wdio-clean-screenshots]' });
-  if (wdioConfig.logLevel !== 'silent' && removedDirs.length > 0) {
-    Logger.log(removedDirs);
-  }
 };
 
 module.exports = cleanSnapshots;