+++ scripts/wdio/getWdioConfigPath.js
@@ -1,5 +1,6 @@
 const path = require('path');
 const fs = require('fs');
+const Logger = require('../utils/logger');
 
 const isFile = filePath => (fs.existsSync(filePath) && !fs.lstatSync(filePath).isDirectory());
 
@@ -8,14 +9,14 @@ const getWdioConfigPath = (configPath) => {
     return path.resolve(configPath);
   }
 
-  // First try to find the local to process.cwd webpack config
+  // Try to find the local to process.cwd webpack config
   const localConfig = path.resolve(process.cwd(), 'wdio.conf.js');
   if (isFile(localConfig)) {
     return localConfig;
   }
 
-  // If that is not found look for the terra-dev-site webpack config.
-  return path.resolve(process.cwd(), 'node_modules', 'terra-dev-site', 'config', 'wdio', 'wdio.conf.js');
+  // Throw error is config not provided and local config is not found.
+  return Logger.error('A wdio config path was not provided and a local config was not found.\n', { context: '[Terra-Toolkit:wdio-runner]:' });
 };
 
 module.exports = getWdioConfigPath;