+++ src/wdio/services/TerraCommands/validate-element.js
@@ -1,40 +1,6 @@
-import accessibilityMethods from './accessiblity';
+import accessibilityMethods from './accessibility';
 import visualRegressionMethods from './visual-regression';
-
-/**
-* Helper method to determine the test name name, the element selector, the mismatch tolerance, and the axe rules
-* @param {Array} args - The list of test arguments to parse.
-*/
-const determineOptions = (...args) => {
-  const param1 = args.length ? args[0] : undefined;
-  const param2 = args.length > 1 ? args[1] : undefined;
-
-  let name = 'default';
-  let options = {};
-  if (typeof param1 === 'string') {
-    name = param1;
-    options = typeof param2 === 'object' && !Array.isArray(param2) ? param2 : options;
-  } else {
-    options = typeof param1 === 'object' && !Array.isArray(param1) ? param1 : options;
-  }
-
-  // Check if custom selector should be used, otherwise use the global value.
-  const selector = options.selector || global.browser.options.terra.selector;
-
-  // Check if custom misMatchTolerance should be used, otherwise use the global value.
-  const misMatchTolerance = options.misMatchTolerance || global.browser.options.visualRegression.compare.misMatchTolerance;
-
-  return {
-    name,
-    selector,
-    misMatchTolerance,
-    axeOptions: {
-      ...options.axeRules && { rules: options.axeRules },
-      restoreScroll: true,
-      context: selector,
-    },
-  };
-};
+import determineOptions from './determine-test-options';
 
 /**
  * A chai assertion method to assert the page is accessible and the screenshot comparison result is within
@@ -51,14 +17,17 @@ const determineOptions = (...args) => {
  */
 const validatesElement = (...args) => {
   const {
+    rules,
+  } = determineOptions.axeOptions(args);
+
+  const {
     name,
     selector,
     misMatchTolerance,
-    axeOptions,
-  } = determineOptions(...args);
+  } = determineOptions.screenshotOptions(args);
 
-  accessibilityMethods.runAccessibilityTest(axeOptions);
-  visualRegressionMethods.runMatchScreenshotTest('withinTolerance', selector, { misMatchTolerance, name });
+  accessibilityMethods.runAccessibilityTest({ rules });
+  visualRegressionMethods.runMatchScreenshotTest(selector, { misMatchTolerance, name });
 };
 
 /**
@@ -73,15 +42,18 @@ const validatesElement = (...args) => {
  */
 const itValidatesElement = (...args) => {
   const {
+    rules,
+  } = determineOptions.axeOptions(args);
+
+  const {
     name,
     selector,
     misMatchTolerance,
-    axeOptions,
-  } = determineOptions(...args);
+  } = determineOptions.screenshotOptions(args);
 
   global.it(`[${name}] is accessible and is within the mismatch tolerance`, () => {
-    accessibilityMethods.runAccessibilityTest(axeOptions);
-    visualRegressionMethods.runMatchScreenshotTest('withinTolerance', selector, { misMatchTolerance });
+    accessibilityMethods.runAccessibilityTest({ rules });
+    visualRegressionMethods.runMatchScreenshotTest(selector, { misMatchTolerance, name });
   });
 };