+++ tests/wdio/beAccessible-spec.js
@@ -1,13 +1,9 @@
-describe('be accessible', () => {
-  const viewports = Terra.viewports('tiny', 'huge');
+const viewports = Terra.viewports('tiny', 'huge');
 
+describe('beAccessible', () => {
   describe('accessible', () => {
     before(() => browser.url('/accessible.html'));
-    Terra.should.beAccessible({ viewports });
     Terra.it.isAccessible({ viewports });
-    it('is accessible', () => {
-      Terra.validates.accessibility({ viewports });
-    });
   });
 
   describe('inaccessible contrast', () => {
@@ -16,10 +12,30 @@ describe('be accessible', () => {
     };
 
     before(() => browser.url('/inaccessible-contrast.html'));
-    Terra.should.beAccessible({ viewports, rules: ignoredA11y });
     Terra.it.isAccessible({ viewports, rules: ignoredA11y });
-    it('is accessible', () => {
-      Terra.validates.accessibility({ viewports, rules: ignoredA11y });
+  });
+});
+
+viewports.forEach((viewport) => {
+  describe('beAccessible with looping', () => {
+    before(() => browser.setViewportSize(viewport));
+
+    describe('accessible', () => {
+      before(() => browser.url('/accessible.html'));
+      it('is accessible', () => {
+        Terra.validates.accessibility();
+      });
+    });
+
+    describe('inaccessible contrast', () => {
+      const ignoredA11y = {
+        'color-contrast': { enabled: false },
+      };
+      before(() => browser.url('/inaccessible-contrast.html'));
+
+      it('is accessible', () => {
+        Terra.validates.accessibility({ rules: ignoredA11y });
+      });
     });
   });
 });