+++ tests/wdio/resize-spec.js
@@ -1,12 +1,12 @@
-Terra.viewports('tiny', 'small', 'large').forEach((viewport) => {
+Terra.viewports().forEach((viewport) => {
   describe('Resize Example', () => {
     before(() => {
       browser.setViewportSize(viewport);
       browser.url('/compare.html');
     });
 
-    Terra.it.isAccessible();
     Terra.it.matchesScreenshot();
+
     it(`resizes ${viewport.name}`, () => {
       const size = browser.getViewportSize();
       expect(size.height).to.equal(viewport.height);